create separate folders for spec files for components and individual functions
